# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Moin Kark API
  version: "1.0.0"
  summary: Read-Only-Aggregator der ChurchDesk-Termine im Kirchenkreis Dithmarschen.
  description: |
    Aggregiert die Termine aller Kirchengemeinde-Organisationen des Kirchenkreises
    Dithmarschen zu einem gemeinsamen Feed und liefert sie als GeoJSON aus.

    Die Schnittstelle ist ausschließlich lesend. Die ChurchDesk-Zugänge bleiben
    server-seitig; Clients kennen nur diese Adresse.

    **Die Antwortformen sind ein Vertrag.** Auf den Geräten laufen App-Versionen
    aus dem Store, die sich nicht mitdeployen lassen — sie lesen genau die hier
    beschriebenen Felder. Aus einem Array wird deshalb kein Objekt, Felder
    verschwinden nicht, Typen ändern sich nicht. Neue Felder dürfen hinzukommen.
    Wer eine Form ändern muss, legt eine neue, versionierte Route an und lässt
    die alte stehen, bis keine App sie mehr ruft.

    Das Zeitfenster ist server-seitig festgelegt (ab heute, Standard 60 Tage) und
    lässt sich nicht per Anfrage steuern. Es beginnt am **Berliner** Kalendertag,
    nicht am UTC-Tag.
# Die dokumentierten Routen sind ohne Anmeldung lesbar — das ist Absicht, nicht
# eine vergessene Absicherung. Geschützt ist allein die Ortsverwaltung (s. unten).
security: []
servers:
  - url: https://api.moin-kark.de
    description: Produktion
  - url: http://localhost:8787
    description: Lokale Entwicklung
...
paths:
...
  /events.geojson:
    get:
      tags: [Feed]
      summary: Alle Termine als GeoJSON
      description: |
        Liefert alle öffentlichen Termine im aktuellen Zeitfenster als
        GeoJSON-FeatureCollection — ein Feature je Termin, dedupliziert über die
        ChurchDesk-Event-ID.

        Nicht öffentliche Kategorien (interne Veranstaltungen, externe
        Raumbuchungen, Kirchengemeinderatssitzungen, Konfirmandenunterricht und
        die über die Ortsverwaltung ausgeschlossenen) fehlen im Feed.

        Antworten werden server-seitig zwischengespeichert. Fällt ChurchDesk
        vorübergehend aus, bleibt der letzte erfolgreiche Stand stehen, statt
        eine leere Sammlung zu liefern. Fällt eine einzelne Organisation aus,
        fehlen ihre Termine — welche das ist, steht in `meta.orgsFailedIds`.

        Die Antwort trägt einen `ETag` (dieselbe Kennung wie `/version.json`);
        mit `If-None-Match` antwortet der Server bei unverändertem Stand mit
        304 ohne Inhalt. Sie darf eine Minute lang zwischengespeichert werden.
      operationId: getEvents
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: Der aktuelle Terminbestand.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventFeatureCollection"
        "304":
          $ref: "#/components/responses/NotModified"
        "500":
          $ref: "#/components/responses/FeedUnavailable"
  /categories.json:
    get:
      tags: [Feed]
      summary: Kategorien des aktuellen Feeds
      description: |
        Die Kategorien aller Termine im Feed, für die Filterauswahl in der App.
        Gleichnamige Kategorien verschiedener Organisationen sind zusammengefasst
        (der Vergleich ignoriert Groß- und Kleinschreibung sowie Randleerzeichen).
        Sortiert nach Häufigkeit, bei Gleichstand alphabetisch.

        Die Antwort ist ein **Array**. `ETag`, `If-None-Match` und
        Zwischenspeichern wie bei `/events.geojson`.
      operationId: getCategories
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: Die Kategorien mit ihrer Häufigkeit.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "304":
          $ref: "#/components/responses/NotModified"
        "500":
          $ref: "#/components/responses/FeedUnavailable"
  /version.json:
    get:
      tags: [Feed]
      summary: Kennung des aktuellen Datenbestands
      description: |
        Eine kurze Kennung über den gesamten Terminbestand. Die App fragt sie
        regelmäßig ab, weil die Antwort wenige Bytes groß ist statt einiger
        hundert Kilobyte. Ändert sich die Kennung, lädt die App den Feed neu.

        Die Kennung ändert sich, sobald sich an einem Termin etwas ändert, das in
        der App sichtbar ist — Titel, Zeiten, Ort, Beschreibung, Bild, Kategorien,
        Preis, Ansprechperson, Gemeindezuordnung oder die Hervorhebung. Eine
        andere Reihenfolge derselben Termine ändert sie nicht.

        Dieselbe Kennung steht im `ETag`; mit `If-None-Match` kommt bei
        unverändertem Stand ein 304 ohne Inhalt.
      operationId: getVersion
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: Kennung und Anzahl der Termine.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
          content:
            application/json:
              schema:
                type: object
                required: [version, count]
                properties:
                  version:
                    type: string
                    description: 16 Hexadezimalzeichen.
                    pattern: "^[0-9a-f]{16}$"
                    examples: ["3f1c9a2b7d4e5061"]
                  count:
                    type: integer
                    description: Anzahl der Termine im Feed.
                    examples: [742]
        "304":
          $ref: "#/components/responses/NotModified"
        "500":
          $ref: "#/components/responses/FeedUnavailable"
...
components:
  parameters:
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: |
        Die zuletzt erhaltene Kennung (`ETag`). Stimmt sie mit dem aktuellen
        Stand überein, antwortet der Server mit 304 ohne Inhalt. Schwache
        (`W/"…"`) und starke Form sowie Listen werden angenommen.
      schema:
        type: string
        examples: ['W/"3f1c9a2b7d4e5061"']
  headers:
    ETag:
      description: |
        Kennung des Datenstands in schwacher Form — dieselbe wie `version` in
        `/version.json`. Schwach, weil die Antwort komprimiert ausgeliefert wird.
      schema:
        type: string
        pattern: '^W/"[0-9a-f]{16}"$'
    CacheControl:
      description: Die Antwort darf eine Minute lang zwischengespeichert werden.
      schema:
        type: string
        const: "public, max-age=60"
  responses:
    NotModified:
      description: Der Stand ist unverändert; die Antwort hat keinen Inhalt.
      headers:
        ETag:
          $ref: "#/components/headers/ETag"
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
    FeedUnavailable:
      description: |
        Es liegt kein Datenstand vor und es konnte auch keiner geladen werden
        (etwa weil alle Organisationen ausgefallen sind). Der Text ist fest;
        die Ursache steht im Server-Log.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            unavailable:
              value: { error: "Datenstand nicht verfügbar." }
  schemas:
    EventFeatureCollection:
      type: object
      required: [type, features]
      properties:
        type:
          type: string
          const: FeatureCollection
        features:
          type: array
          description: Ein Eintrag je Termin. Bei leerem Bestand ein leeres Array.
          items:
            $ref: "#/components/schemas/EventFeature"
        meta:
          $ref: "#/components/schemas/FeedMeta"
    EventFeature:
      type: object
      required: [type, geometry, properties]
      properties:
        type:
          type: string
          const: Feature
        geometry:
          $ref: "#/components/schemas/Point"
        properties:
          $ref: "#/components/schemas/EventProps"
    Point:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          const: Point
        coordinates:
          type: array
          description: |
            GeoJSON-Reihenfolge: **[Länge, Breite]**, nicht umgekehrt.
            Der Nullpunkt 0/0 kommt nicht vor — ChurchDesk liefert ihn für nicht
            verortete Adressen, dort greift stattdessen ein Gemeindepunkt.
          minItems: 2
          maxItems: 2
          items:
            type: number
          examples: [[8.861245, 54.129605]]
...
    Category:
      type: object
      description: Ein Eintrag der Kategorienliste für die Filterauswahl.
      required: [title, color, count]
      properties:
        title:
          type: string
          examples: ["Gottesdienst"]
        color:
          type: integer
        count:
          type: integer
          description: Anzahl der Termine mit dieser Kategorie im aktuellen Feed.
    FeedMeta:
      type: object
      description: |
        Angaben zum Datenstand, für Betrieb und Fehlersuche. Nicht Teil des
        GeoJSON-Standards.
      required:
        - generatedAt
        - from
        - to
        - windowFrom
        - windowTo
        - total
        - withEventCoords
        - withFallbackCoords
        - orgsOk
        - orgsFailed
        - orgsFailedIds
        - orgsConfigured
        - orgsMissing
      properties:
        generatedAt:
          type: string
          format: date-time
          description: Wann dieser Stand erzeugt wurde.
        from:
          type: string
          format: date-time
          description: |
            Zeitpunkt, zu dem dieser Stand angefordert wurde — **nicht** der
            Beginn des Zeitfensters. Das Fenster beginnt am Berliner
            Kalendertag dieses Zeitpunkts, siehe `windowFrom`: Termine von
            heute früh sind enthalten, auch wenn `from` später am Tag liegt.
        to:
          type: string
          format: date-time
          description: |
            Ein Zeitpunkt am letzten Tag des Zeitfensters (Mittag UTC). Der
            Tag selbst steht in `windowTo`.
        windowFrom:
          type: string
          format: date
          description: |
            Erster Kalendertag des Zeitfensters in Europe/Berlin. Genau dieses
            Datum bekommt ChurchDesk als Beginn.
          examples: ["2026-09-14"]
        windowTo:
          type: string
          format: date
          description: Letzter Kalendertag des Zeitfensters in Europe/Berlin.
          examples: ["2026-11-13"]
        total:
          type: integer
          description: Anzahl der Termine.
        withEventCoords:
          type: integer
          description: Davon mit eigener Koordinate aus ChurchDesk.
        withFallbackCoords:
          type: integer
          description: Davon auf einem Gemeinde- oder Organisationspunkt.
        orgsOk:
          type: integer
          description: Organisationen, die geantwortet haben.
        orgsFailed:
          type: integer
          description: |
            Organisationen, die ausgefallen sind. Deren Termine fehlen im Feed,
            ohne dass die Antwort sonst darauf hinweist — welche es sind, steht
            in `orgsFailedIds`.
        orgsFailedIds:
          type: array
          description: |
            Die `orgId`s der ausgefallenen Organisationen (gleiche Kennung wie
            `orgId` in den Termin-Eigenschaften). Leer, wenn alle geantwortet
            haben. Die App darf gemerkte Termine dieser Organisationen nicht
            als abgesagt behandeln, nur weil sie im Feed fehlen.
          items:
            type: integer
          examples: [[2725]]
        orgsConfigured:
          type: integer
          description: |
            Organisationen mit hinterlegtem Zugang — nur die werden abgefragt
            (`orgsOk` + `orgsFailed`).
        orgsMissing:
          type: integer
          description: |
            Organisationen **ohne** hinterlegten Zugang. Sie werden gar nicht
            abgefragt; ihre Termine fehlen dauerhaft, ohne je als Ausfall zu
            zählen. Im Normalbetrieb 0.
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Beschreibung des Fehlers.

# Nicht dokumentiert: die Ortsverwaltung unter /admin.
#
# Dort werden Koordinaten-Korrekturen, Kategorie-Ausschlüsse und
# Hervorhebungen zur Laufzeit gepflegt. Die Oberfläche liegt unter /admin, die
# zugehörigen Routen unter /admin/api/. Sie sind durch ein Token geschützt
# (mindestens 24 Zeichen, sonst bleibt die Verwaltung abgeschaltet) und ohne
# gesetztes Token abgeschaltet; nach zehn Fehlversuchen von einer Adresse
# innerhalb einer Minute antworten sie mit 429. Sie gehören nicht zur
# öffentlichen Schnittstelle und keine App ruft sie.